requirements.txt
click==8.1.3
colorama==0.4.5
coloredlogs==15.0.1
cssutils==2.5.1
dotmap==1.3.30
pyyaml==6.0
isort==5.10.1
prospector==1.7.7
black==22.6.0
bandit==1.7.4
mypy==0.971
pyroma==4.0
types-PyYAML==6.0.11
pylint==2.14.5
